AddGasto corrregido

+++ frontend/src/utils/FomularioDinamico.jsx
@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { eliminarTelefonos, eliminarTitulos } from '../api/gestionAcademicos';
 import { eliminarCuentasBancarias } from '../api/gestionProveedores';
+import { agregarProductoServicio } from '../api/gestionGastos'; // Importar la función
 import Accordion from '@mui/material/Accordion';
 import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
@@ -9,6 +10,8 @@ import Typography from '@mui/material/Typography';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 
 export const FormularioDinamico = ({ items, setItems, configuracion, itemName }) => {
+    const [productoNuevo, setProductoNuevo] = useState({});
+
     const agregarItem = () => {
         const nuevoItem = {};
         configuracion.forEach(conf => {
@@ -30,19 +33,28 @@ export const FormularioDinamico = ({ items, setItems, configuracion, itemName })
         setItems(nuevosItems);
     };
 
-    const eliminarItem = (index) => {
-        const nuevosItems = [...items];
-        nuevosItems.splice(index, 1);
-        if (Object.keys(items[0])[0] === "id_telefono") {
-            eliminarTelefonos(items[index].id_telefono, localStorage.getItem("token"));
+    const eliminarItem = async (index) => {
+        try {
+            if (itemName === "Producto Servicio") {
+                await agregarProductoServicio(items[index], localStorage.getItem("token")); // Llamar a la función del API para guardar
+            } else {
+                if (Object.keys(items[0])[0] === "id_telefono") {
+                    eliminarTelefonos(items[index].id_telefono, localStorage.getItem("token"));
+                }
+                if (Object.keys(items[0])[0] === "id_titulos") {
+                    eliminarTitulos(items[index].id_titulos, localStorage.getItem("token"));
+                }
+                if (Object.keys(items[0])[0] === "id_numero") {
+                    eliminarCuentasBancarias(items[index].id_numero, localStorage.getItem("token"));
+                }
+            }
+            const nuevosItems = [...items];
+            nuevosItems.splice(index, 1);
+            setItems(nuevosItems);
+        } catch (error) {
+            console.error('Error al agregar producto/servicio:', error);
+            // Manejar el error apropiadamente (mensaje de error, notificación, etc.)
         }
-        if (Object.keys(items[0])[0] === "id_titulos") {
-            eliminarTitulos(items[index].id_titulos, localStorage.getItem("token"));
-        }
-        if (Object.keys(items[0])[0] === "id_numero") {
-            eliminarCuentasBancarias(items[index].id_numero, localStorage.getItem("token"));
-        }
-        setItems(nuevosItems);
     };
 
     const dividirEnGruposDeDos = (array) => {
@@ -57,16 +69,14 @@ export const FormularioDinamico = ({ items, setItems, configuracion, itemName })
 
     return (
         <>
-
             {items.map((item, index) => (
-                <Accordion>
+                <Accordion key={index}>
                     <AccordionSummary expandIcon={<ExpandMoreIcon />} aria-controls="panel1a-content" id="panel1a-header">
                         <Typography>{itemName}  {index + 1}</Typography>
                     </AccordionSummary>
                     <AccordionDetails>
-
-                        {grupos.map((grupo) => (
-                            <div key={index} className="row mb-4">
+                        {grupos.map((grupo, grupoIndex) => (
+                            <div key={grupoIndex} className="row mb-4">
                                 {grupo.map(conf => (
                                     <div key={conf.campo} className="col-md-6">
                                         <label htmlFor={conf.campo} className="label-personalizado mb-2">{conf.placeholder}</label>
@@ -98,13 +108,15 @@ export const FormularioDinamico = ({ items, setItems, configuracion, itemName })
                                 ))}
                             </div>
                         ))}
-
-
-                        <button type="button" className="mb-2 eliminarBtn" onClick={() => eliminarItem(index)}>Eliminar</button>
+                        {itemName === "Producto Servicio" ? (
+                            <button type="button" className="mb-2 eliminarBtn" onClick={() => eliminarItem(index)}>Guardar</button>
+                        ) : (
+                            <button type="button" className="mb-2 eliminarBtn" onClick={() => eliminarItem(index)}>Eliminar</button>
+                        )}
                     </AccordionDetails>
                 </Accordion>
             ))}
-            <button type="button" className={items.length == 0 ? "agregarBtn": "agregarBtnMod"} style={{ width: "100%" }} onClick={agregarItem}>+</button>
+            <button type="button" className={items.length === 0 ? "agregarBtn" : "agregarBtnMod"} style={{ width: "100%" }} onClick={agregarItem}>+</button>
         </>
     );
 };
